+++ moshi-ir/moshi-gradle-plugin/src/main/kotlin/dev/zacsweers/moshix/ir/gradle/MoshiGradleSubplugin.kt
@@ -2,16 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.moshix.ir.gradle
 
-import com.google.devtools.ksp.gradle.KspExtension
 import dev.zacsweers.moshi.ir.gradle.VERSION
-import java.util.Locale.US
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.UnknownTaskException
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.TaskProvider
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.dsl.kotlinExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilerPluginSupportPlugin
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
@@ -20,55 +16,13 @@ import org.jetbrains.kotlin.gradle.plugin.SubpluginOption
 
 class MoshiGradleSubplugin : KotlinCompilerPluginSupportPlugin {
 
-  private companion object {
+  internal companion object {
     val SUPPORTED_PLATFORMS = setOf(KotlinPlatformType.androidJvm, KotlinPlatformType.jvm)
   }
 
   override fun apply(target: Project) {
-    target.extensions.create("moshi", MoshiPluginExtension::class.java)
-    if (
-      target.providers.gradleProperty("moshix.generateProguardRules").orNull?.toBoolean() != false
-    ) {
-      try {
-        target.pluginManager.apply("com.google.devtools.ksp")
-      } catch (e: Exception) {
-        // KSP not on the classpath, ask them to add it
-        error(
-          "MoshiX proguard rule generation requires KSP to be applied to the project. " +
-            "Please apply the KSP Gradle plugin ('com.google.devtools.ksp') to your buildscript and try again."
-        )
-      }
-
-      fun addKspDep(configurationName: String) {
-        target.dependencies.add(
-          configurationName,
-          "dev.zacsweers.moshix:moshi-proguard-rule-gen:$VERSION",
-        )
-      }
-
-      // Add the KSP dependency to the appropriate configurations
-      // In KMP, we only add to androidJvm/jvm targets
-      val kExtension = target.kotlinExtension
-      if (kExtension is KotlinMultiplatformExtension) {
-        kExtension.targets
-          .matching { it.platformType in SUPPORTED_PLATFORMS }
-          .configureEach {
-            addKspDep(
-              "ksp${it.targetName.replaceFirstChar { if (it.isLowerCase()) it.titlecase(US) else it.toString() }}"
-            )
-          }
-      } else {
-        addKspDep("ksp")
-      }
-      target.extensions.configure(KspExtension::class.java) {
-        // Enable core moshi proguard rule gen
-        it.arg("moshi.generateCoreMoshiProguardRules", "true")
-        // Disable moshi's KSP codegen, we're doing it ourselves
-        it.excludeProcessor(
-          "com.squareup.moshi.kotlin.codegen.ksp.JsonClassSymbolProcessorProvider"
-        )
-      }
-    }
+    val extension = target.extensions.create("moshi", MoshiPluginExtension::class.java)
+    target.afterEvaluate { extension.apply(target) }
   }
 
   override fun getCompilerPluginId(): String = "dev.zacsweers.moshix.compiler"
