--- moshi-ir/moshi-kotlin-tests/gradle.properties
@@ -1,1 +1,0 @@
-moshix.generateProguardRules=true
